refactor: move useDeploy to deploy context
