+++ src/components/Deploy/RestartFlied/index.tsx
@@ -1,4 +1,4 @@
-import { useDeployContext } from 'src/contexts/deployContext'
+import { useDeployContext } from 'src/contexts'
 import { FieldHeader } from '../FieldHeader'
 import { DeployActionEnum } from 'src/types/deployTypes'
