+++ src/contexts/DeployContext.tsx
@@ -6,6 +6,10 @@ import { useReducer } from 'react'
 import { DeployMethodsEnum } from 'src/types/deployTypes'
 import { deployReducer } from 'src/utils/deployReducer'
 
+import { createContext, useContext } from 'react'
+import { DeployActions } from 'src/utils/deployReducer'
+import { DeployStateType } from '../types/deployTypes'
+
 const initialState = {
   operations: [],
   filterExpression: '',
@@ -14,7 +18,13 @@ const initialState = {
   restart: false,
 }
 
-export function useDeploy() {
+const DeployContext = createContext<{
+  deploy: () => Promise<void>
+  state: DeployStateType
+  dispatch: React.Dispatch<DeployActions>
+}>(null)
+
+export function DeployContextProvider({ children }) {
   const { deployApi } = useTdpClient()
   const [state, dispatch] = useReducer(deployReducer, initialState)
 
@@ -31,5 +41,19 @@ export function useDeploy() {
     res?.data?.state && toast.info(`Deploy status: ${res.data.state}`)
   }
 
-  return { deploy, state, dispatch }
+  return (
+    <DeployContext.Provider value={{ deploy, state, dispatch }}>
+      {children}
+    </DeployContext.Provider>
+  )
+}
+
+export function useDeployContext() {
+  const deployContext = useContext(DeployContext)
+  if (!deployContext)
+    throw new Error(
+      'useDeployContext() hook must be inside a DeployContextProvider'
+    )
+
+  return deployContext
 }
