+++ src/contexts/index.ts
@@ -1,2 +1,3 @@
 export * from './AuthContextProvider'
+export * from './DeployContext'
 export * from './TdpClientContext'
