+++ src/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from './useComponentInfos'
-export * from './useDeploy'
 export * from './useServerStatus'
 export * from './useServiceInfos'
 export * from './useServicesList'
